Crate parquet_layout

Binary that prints the physical layout of a parquet file

Install

parquet-layout can be installed using cargo:

cargo install parquet --features=cli

After this parquet-layout should be available:

parquet-layout XYZ.parquet

The binary can also be built from the source code and run as follows:

cargo run --features=cli --bin parquet-layout XYZ.parquet

Structs

Args ğŸ”’

ColumnChunk ğŸ”’

Page ğŸ”’

ParquetFile ğŸ”’

RowGroup ğŸ”’

Functions

compression ğŸ”’

Returns a string representation for a given compression

do_layout ğŸ”’

encoding ğŸ”’

Returns a string representation for a given encoding

main ğŸ”’

read_page_header ğŸ”’

Reads the page header at offset from reader, returning both the PageHeader and its length in bytes